Move sentry.ts into api directory
