+++ site/source/entries/entry-en.tsx
@@ -8,7 +8,7 @@ import ruleTranslations from '../locales/rules-en.yaml'
 import translateRules from '../locales/translateRules'
 import translations from '../locales/ui-en.yaml'
 
-import '../sentry'
+import '../api/sentry'
 
 export const AppEn = () => (
 	<I18nProvider locale="en-GB">
